+++ {{cookiecutter.github_repository_name}}/{{cookiecutter.app_name}}/users/admin.py
@@ -1,35 +1,6 @@
-from __future__ import absolute_import, unicode_literals
-
-from django import forms
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as AuthUserAdmin
-from django.contrib.auth.forms import UserChangeForm, UserCreationForm
-
-from .models import User
-
-
-class CustomUserChangeForm(UserChangeForm):
-
-    class Meta(UserChangeForm.Meta):
-        model = User
-
-
-class CustomUserCreationForm(UserCreationForm):
-
-    # http://james.lin.net.nz/2013/06/08/django-custom-user-model-in-admin-relation-auth_user-does-not-exist/
-    def clean_username(self):
-        # Since User.username is unique, this check is redundant,
-        # but it sets a nicer error message than the ORM. See #13147.
-        username = self.cleaned_data["username"]
-        try:
-            User._default_manager.get(username=username)
-        except User.DoesNotExist:
-            return username
-        raise forms.ValidationError(self.error_messages['duplicate_username'])
-
-    class Meta(UserCreationForm.Meta):
-        model = User
-
+from .forms import CustomUserChangeForm, CustomUserCreationForm
 
 @admin.register(User)
 class UserAdmin(AuthUserAdmin):
